:recycle: refactor: Modifying the pafs of the funcs
